+++ src/apis/products.ts
@@ -1,6 +1,6 @@
 import type { Product } from '../types/product';
 
-const URL = '/products';
+const URL = '/products/empty';
 
 export const fetchProducts = async () => {
   const response = await fetch(URL);
